Plot method for BRIL objects

Plot method for BRIL objects

# S3 method for BRIL
plot(
  x,
  contents = "plot",
  showClusters = TRUE,
  showMode = TRUE,
  col,
  colMode,
  colClusters = NULL,
  iterationsIndices = NULL,
  iterationsOptions = NULL,
  ...
)

Arguments

x
An object of class BRIL (see bril())

contents
Contents to be displayed, options are "scatterplot", or "iterations" (only one option possible)

showClusters
Logical value. Used with contents="scatterplot", to show or not the different clusters

showMode
Logical value. Used for contents="clusters", to show or not the main mode to show or not the different clusters

col
default color of samples

colMode
Color of the mode when contents="scatterplot" (if showMode is TRUE)

colClusters
List (or array) of colors (length must be at least equal to the number of clusters identified by the function bril(), i.e. `x$nbClusters)

iterationsIndices
Numerical value or array of numerical values, used with contents="iterations", providing the indices of the iterations to be ploted. If more than one iteration is requested, an interactive menu in the console will be used for the selection. 0 or NULL (default) will include all the iterations. Values that are negative or superior to the number of iterations performed by the execution bril() will be ignored

iterationsOptions
List of additional parameters to be passed to the plot.BRIL.Filtering function when contents="iterations" is selected (see plot.BRIL.Filtering for details). Example: iterationsOptions = list(Xlab=NA, Ylab=NA, contents=c("p.values", "scatterplot"))

...
Other arguments passed to or from other methods (such as pch for the symbols, main and sub for title and subtitle, xlab, xmin, ...)

See also